% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf_clone.R
\name{taf_clone}
\alias{taf_clone}
\title{Perform a clone from the ICES TAF GitHub organisation}
\usage{
taf_clone(repo_name, local_dir = NULL)
}
\arguments{
\item{repo_name}{the name of the TAF repo you want to clone}

\item{local_dir}{the local directory you want to clone into}
}
\value{
a git2r repository object
}
\description{
user must have approprate access privileges
}
