% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf_test.R
\name{taf_test}
\alias{taf_test}
\title{Perform a test run of a TAF repository}
\usage{
taf_test(repo_name, local_dir = NULL, overwrite = FALSE)
}
\arguments{
\item{repo_name}{the name of the TAF repo you want to clone}

\item{local_dir}{the local directory you want to clone into}

\item{overwrite}{if local_dir exists, should it be overwritten}
}
\value{
a git2r repository object
}
\description{
user must have approprate access privileges
}
\examples{

\dontrun{
taf_test(
  "2019_TAF_template",
  local_dir = "mytest",
  overwrite = TRUE
)
}

}
