% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.ggplot.R
\name{zoom.ggplot}
\alias{zoom.ggplot}
\title{Zoom ggplot}
\usage{
\method{zoom}{ggplot}(x, size = 28, main = 1.2 * size, axis = 1.3 *
  size, adjust = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{"ggplot"}.}

\item{size}{text size multiplier.}

\item{main}{size of main title (default is \code{1.2 * size}).}

\item{axis}{size of tick labels (default is \code{size}).}

\item{adjust}{whether to center plot title.}

\item{\dots}{further arguments, currently ignored.}
}
\value{
The same ggplot object, but with altered text size.
}
\description{
Change text size in a ggplot.
}
\note{
The default values result in ggplots that have similar text size as base
plots, when using \code{taf.png}.

This function ends with a \code{\link[=print.ggplot]{print}} call, to make it
easy to export the ggplot to a file, without the need of an explicit
\code{print}.
}
\examples{
library(icesTAF)
library(ggplot2)

qplot(1,1)
zoom(qplot(1,1))
zoom(qplot(1,1), size=1, axis=0.8)

\dontrun{
taf.png("myplot")
plot(1)
dev.off()

taf.png("myggplot")
zoom(qplot(1,1))
dev.off()
}

}
\seealso{
\code{\link[ggplot2]{ggplot}} initializes a ggplot.

\code{\link[icesTAF]{taf.png}} opens a PNG graphics device.
}
